requirements.txt
cereal-lazer==0.1.1
flask-restless==0.17.0
pbr>=3.0
